Remove script antigo da ficha clínica
