+++ pages/funcionarios/vet-ficha-clinica.html
@@ -340,7 +340,7 @@
     <script src="../../scripts/core/ui.js"></script>
     <script src="../../scripts/core/main.js"></script>
     <script src="../../scripts/funcionarios/funcionarios.js"></script>
-    <script src="../../scripts/funcionarios/vet/ficha-clinica.js"></script>
+    <script type="module" src="../../scripts/funcionarios/vet/ficha-clinica/index.js"></script>
 </body>
 
 </html>
