+++ scripts/funcionarios/vet/ficha-clinica/index.js
@@ -1,0 +1,6 @@
+// Main entry for the Vet ficha clínica page
+import { initFichaClinica } from './init.js';
+
+document.addEventListener('DOMContentLoaded', () => {
+  initFichaClinica();
+});
